Changed all damage types to int instead of float

+++ src/com/java/Axe.java
@@ -5,7 +5,7 @@ public class Axe extends Weapon {
     private String[] validSubtypes = {"rusty axe", "battleaxe", "felling axe", "halberd", "the labrys"};
     protected String subType; 
 
-    public Axe(String name, float dmg, String subType, int speedPenalty) {
+    public Axe(String name, int dmg, String subType, int speedPenalty) {
         super(name, dmg, "axe", speedPenalty);
         for(int i = 0; i < validSubtypes.length; i++) {
             if(subType.equals(validSubtypes[i])) {
