+++ src/com/java/Dagger.java
@@ -5,7 +5,7 @@ public class Dagger extends Weapon {
     private String[] validSubtypes = {"rusty dagger", "rondel dagger", "stiletto dagger", "castillon dagger", "sacrificial dagger"};
     protected String subType; 
 
-    public Dagger(String name, float dmg, String subType, int speedPenalty) {
+    public Dagger(String name, int dmg, String subType, int speedPenalty) {
         super(name, dmg, "dagger", speedPenalty);
         for(int i = 0; i < validSubtypes.length; i++) {
             if(subType.equals(validSubtypes[i])) {
